+++ tests/test_messages.py
@@ -6,6 +6,7 @@ import pytest
 from urlobject import URLObject
 from nylas.client.restful_models import Message
 from nylas.utils import timestamp_from_dt
+import pytz
 
 
 @pytest.mark.usefixtures("mock_messages")
@@ -119,6 +120,17 @@ def test_slice_messages(api_client):
 @pytest.mark.usefixtures("mock_messages")
 def test_filter_messages_dt(mocked_responses, api_client):
     api_client.messages.where(received_before=datetime(2010, 6, 1)).all()
+    assert len(mocked_responses.calls) == 1
+    request = mocked_responses.calls[0].request
+    url = URLObject(request.url)
+    assert url.query_dict["received_before"] == "1275350400"
+
+
+@pytest.mark.usefixtures("mock_messages")
+def test_filter_messages_dt_with_timezone(mocked_responses, api_client):
+    api_client.messages.where(
+        received_before=datetime(2010, 6, 1, tzinfo=pytz.utc)
+    ).all()
     assert len(mocked_responses.calls) == 1
     request = mocked_responses.calls[0].request
     url = URLObject(request.url)