+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased (dev)
 * Add `metadata` field in the Event model to support new event metadata feature
 * Add filtering support for `metadata_pair`
 * Fix adding a tracking object to an existing `draft`
+* Fix issue when converting offset-aware `datetime` objects to `timestamp`
 
 v4.12.1
 -------